pyproject.toml
[project]
name = "openenv-support-agent"
version = "0.1.0"
description = "Support Agent Environment for OpenEnv"
authors = [
  { name = "Jatin Chaudhary" }
]

requires-python = ">=3.9"

dependencies = [
    "openenv-core[core]==0.2.3",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "huggingface-hub>=0.30.0"
]
[project.scripts]
server = "server.app:main"

[tool.openenv]
entrypoint = "server.app:main"